chore: rename waterfall chart files (phase 2)
